backend_flask/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography>=42.0.0
PyJWT==2.10.1
bcrypt==4.2.1
pytest==8.3.4
